# Speed up PK lookups at index time.