@jpountz 2016-08-09
This is basically a revert of #14070. I'm stalling this on making cache keys / closed listeners less trappy.